ponderada_8/server/requirements.txt
flask
openai < 1
python-dotenv